Move resource caches
