+++ src/Abstractions/NexusMods.Abstractions.Resources.Caching/NexusMods.Abstractions.Resources.Caching.csproj
@@ -1,0 +1,14 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <!-- NuGet Package Shared Details -->
+    <Import Project="$([MSBuild]::GetPathOfFileAbove('NuGet.Build.props', '$(MSBuildThisFileDirectory)../'))" />
+
+    <ItemGroup>
+        <ProjectReference Include="..\NexusMods.Abstractions.Resources\NexusMods.Abstractions.Resources.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <PackageReference Include="BitFaster.Caching" />
+    </ItemGroup>
+
+</Project>
